backend/eduslide-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-pptx==0.6.23
PyPDF2==3.0.1
pdfplumber==0.10.3
anthropic==0.7.8
python-multipart==0.0.6
pillow==10.1.0
python-dotenv==1.0.0
aiofiles==23.2.1